requirements.txt
httpx>=0.27.0
pydantic>=2.7.0
python-dotenv>=1.0.1
pypdf>=4.2.0
pymupdf>=1.24.0
pillow>=10.4.0
pytesseract>=0.3.10
transformers>=4.36.0
torch>=2.1.0
accelerate>=0.25.0

